+++ src/Settings/LANGUAGE_SETTINGS_KEY.mjs
@@ -1,1 +1,0 @@
-export const LANGUAGE_SETTINGS_KEY = "language";